fix wrong goal orientation
Problem:
The robot doesn't always stop at the goal orientation

Steps to reproduce:
* Set a goal position where the robot has to move to have the right orientation

Cause:
The robot rotates at `rotate_to_heading_angular_vel` and starts to slow when reaching the wanted orientation, but the orientation is validated as reached.

Solution:
Add a parameter specific for angular velocity when rotating to goal heading. This parameter should be slower so the rbot has the time to stop at the heading